indexer/requirements.txt
chromadb==0.6.3
paho-mqtt==2.1.0
prometheus-client==0.21.1
psycopg[binary]==3.2.6
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
